0.9.8.7 1. расчленение класса viewer 2. создан класс viewer_processor, который отвечает за данные перед отрисовкой в классе viewer 3. viewer_processor работает, пока, только с TXT файлами

+++ viewer_widget/viewer/pix_filter_panel.cpp
@@ -149,7 +149,7 @@ QColor Pix_Filter_Panel::getPenColor() const
     return ui->edit_panel->getPenColor();
 }
 
-int Pix_Filter_Panel::getPenValue() const
+double Pix_Filter_Panel::getPenValue() const
 {
     return ui->edit_panel->getPenValue();
 }
